+++ CHANGELOG.md
@@ -17,5 +17,10 @@
 
 - Server pages should be in folder `~/public_html/` now
   (Use `./bin/create_server_pages "${HOME}" "${WEB_USER}"` to create required links, Eg. for Raspberry Pi install: `./bin/create_server_pages /home/pi web`)
-- Config file name changed to susanoo_WeatherStation.conf
-  (Use `mv ~/.config/meteo.conf ~/.config/susanoo_WeatherStation.conf` to rename it)
+- Config file name changed to susanoo_WeatherStation.conf . Use the below to rename it:
+
+      mv ~/.config/meteo.conf ~/.config/susanoo_WeatherStation.conf
+      sudo ln -f -s "${HOME}/.config/susanoo_WeatherStation.conf" "${HOME}/../${WEB_USER}/.config/susanoo_WeatherStation.conf"
+      chmod +x "${HOME}/.config"
+      sudo chown "${WEB_USER}" "${HOME}/../${WEB_USER}/.config/susanoo_WeatherStation.conf"
+      sudo chgrp "${WEB_USER}" "${HOME}/../${WEB_USER}/.config/susanoo_WeatherStation.conf"